Move assembly documentation files to /lib/asm-docs
