+++ .eslintignore
@@ -7,5 +7,5 @@ out
 views
 
 # Autogenerated files
-lib/asm-providers/generated/asm-docs-*.js
+lib/asm-docs/generated/asm-docs-*.js
 etc/scripts/vendor/jvms.html
