+++ .nycrc.yml
@@ -3,7 +3,7 @@ include:
   - lib/**/*.js
   - lib/**/*.ts
 exclude:
-  - lib/asm-providers/generated/asm-docs-*.js
+  - lib/asm-docs/generated/asm-docs-*.js
   - lib/compilers/fake-for-test.js
 report-dir: ./out/coverage
 reporter:
