+++ .prettierignore
@@ -40,7 +40,7 @@ out
 views
 
 # Autogenerated files
-lib/asm-providers/generated/asm-docs-*.js
+lib/asm-docs/generated/asm-docs-*.js
 
 
 #########################
